[pytx] Initial rename
