+++ python-threatexchange/benchmarks/benchmark_pdq_faiss_matchers.py
@@ -9,9 +9,9 @@ import pickle
 import numpy
 import faiss
 
-from threatexchange.hashing.pdq_utils import BITS_IN_PDQ
+from threatexchange.signal_type.pdq.pdq_utils import BITS_IN_PDQ
 
-from threatexchange.hashing.pdq_faiss_matcher import (
+from threatexchange.signal_type.pdq.pdq_faiss_matcher import (
     PDQFlatHashIndex,
     PDQMultiHashIndex,
 )
