+++ python-threatexchange/threatexchange/exchanges/impl/ncmec_api.py
@@ -21,7 +21,7 @@ from threatexchange.exchanges.collab_config import (
 )
 from threatexchange.signal_type.signal_base import SignalType
 from threatexchange.signal_type.md5 import VideoMD5Signal
-from threatexchange.signal_type.pdq import PdqSignal
+from threatexchange.signal_type.pdq.pdq import PdqSignal
 
 
 _API_NAME: str = "ncmec"
