+++ python-threatexchange/threatexchange/extensions/pdq_ocr/pdq_ocr.py
@@ -11,14 +11,14 @@ import pathlib
 from threatexchange.content_type.content_base import ContentType
 from threatexchange.content_type.photo import PhotoContent
 
-from threatexchange.signal_type.pdq import PdqSignal
+from threatexchange.signal_type.pdq.pdq import PdqSignal
 from threatexchange.signal_type.raw_text import RawTextSignal
 from threatexchange.signal_type import signal_base
 from threatexchange.exchanges.impl.fb_threatexchange_signal import (
     HasFbThreatExchangeIndicatorType,
 )
 
-from threatexchange.hashing.pdq_hasher import pdq_from_file
+from threatexchange.signal_type.pdq.pdq_hasher import pdq_from_file
 from threatexchange.extensions.pdq_ocr.ocr_utils import text_from_image_file
